refact : schedule
