+++ job-rocket-backend/src/main/java/rocket/jobrocketbackend/schedule/dto/ScheduleCreateDTO.java
@@ -21,20 +21,5 @@ public class ScheduleCreateDTO {
     private LocalDate dueDate;
     private String memo;
     private String state;
-    private Long userId;
 
-    public ScheduleEntity toCreateEntity(UserEntity user){
-        return ScheduleEntity.builder()
-                .title(this.getTitle())
-                .type(ScheduleType.Document)
-                .dueDate(this.getDueDate())
-                .memo(this.getMemo())
-                .user(user)
-                .state(ScheduleState.from(this.getState()))
-                .build();
-    }
-
-    public void setUserId(Long userId){
-        this.userId = userId;
-    }
 }
