+++ job-rocket-backend/src/test/java/rocket/jobrocketbackend/schedule/controller/ScheduleControllerTest.java
@@ -22,6 +22,7 @@ import rocket.jobrocketbackend.schedule.service.ScheduleService;
 import java.time.LocalDate;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
@@ -95,7 +96,7 @@ class ScheduleControllerTest {
                 .title("제목")
                 .state(ScheduleState.Ongoing)
                 .build();
-        when(scheduleService.create(any(ScheduleCreateDTO.class))).thenReturn(result);
+        when(scheduleService.create(any(ScheduleCreateDTO.class),anyLong())).thenReturn(result);
         //when
         //then
         mockMvc.perform(
