+++ job-rocket-backend/src/test/java/rocket/jobrocketbackend/schedule/dto/ScheduleDTOTest.java
@@ -27,7 +27,7 @@ class ScheduleDTOTest {
                 .dueDate(LocalDate.of(2024, 12, 23))
                 .build();
         //when
-        ScheduleEntity newScheduleEntity = dto.toCreateEntity(new UserEntity());
+        ScheduleEntity newScheduleEntity = ScheduleEntity.create(dto,new UserEntity());
         //then
         assertThat(newScheduleEntity.getTitle()).isEqualTo("제목입니다");
         assertThat(newScheduleEntity.getMemo()).isEqualTo("메모입니다");
@@ -35,32 +35,7 @@ class ScheduleDTOTest {
         assertThat(newScheduleEntity.getType()).isEqualTo(ScheduleType.Document);
         assertThat(newScheduleEntity.getState()).isEqualTo(ScheduleState.Ongoing);
     }
-    @Test
-    @DisplayName("ScheduleCreateDTO에 유효하지 않은 state를 넣으면 예외를 발생시킨다.")
-    void toNewScheduleEntityThrowException() {
-        //given
-        ScheduleCreateDTO dto = ScheduleCreateDTO.builder().memo("메모입니다")
-                .title("제목입니다")
-                .state("ongoing")
-                .dueDate(LocalDate.of(2024, 12, 23))
-                .build();
-        //when
-        //then
-        assertThatThrownBy(() -> dto.toCreateEntity(new UserEntity()))
-                .isInstanceOf(IllegalScheduleStateException.class)
-                .hasMessage("잘못된 상태 값입니다.");
-    }
-    @Test
-    @DisplayName("ScheduleType에 유효하지 않은 state를 넣으면 예외를 발생시킨다.")
-    void toNewScheduleEntityTypeThrowException() {
-        //given
-        ScheduleModifyTypeRequest request = ScheduleModifyTypeRequest.builder().scheduleId(1L).type("3차면접").build();
-        //when
-        //then
-        assertThatThrownBy(request::toTypeModifyDto)
-                .isInstanceOf(IllegalScheduleTypeException.class)
-                .hasMessage("잘못된 타입 값입니다.");
-    }
+
     @Test
     @DisplayName("ScheduleEntity를 ScheduleDTO로 변환")
     void scheduleDtoFromScheduleEntity(){
